Daily_Regression_Test/tele_op_services/requirements.txt
pytest>=8.0.0
requests>=2.31.0
allure-pytest>=2.13.0
python-socketio[client]==5.11.3
